metaflow/requirements.txt
# Core dependencies
pyspark>=3.3.0,<4.0.0
delta-spark>=2.3.0
sqlalchemy>=1.4.0,<2.0.0
pyyaml>=6.0
click>=8.0.0
python-dotenv>=0.19.0
jinja2>=3.0.0

# Database drivers (optional, installed as extras)
# psycopg2-binary>=2.9.0  # PostgreSQL
# pymysql>=1.0.0          # MySQL
# cx_Oracle>=8.3.0        # Oracle

# Kafka support
kafka-python>=2.0.2

# Logging and monitoring
rich>=12.0.0
structlog>=21.5.0

# Configuration validation
pydantic>=1.9.0